Fix casing of resource directory names.
The casing of resource directories must match the casing of the associated class. Otherwise no UI elements are shown in Jenkins’ settings page (works in Jenkins 1 though).

This means the resource directory of a class can no longer contain a Messages.properties file. Otherwise the Messages.class is generated in a package that matches the class name.

This fixes #2.