No longer auto-load the Context class
If the Context is automatically instantiated this causes two injectors to be created, which in turn creates two instances of classes that should only have a single instance. E.g. the SQS trigger’s descriptor will receive one instance of the SQS queue monitor scheduler, while actual monitor instances will be registered with a second instance of the scheduler. This in turn causes reconfiguration of the scheduler to “fail” because the wrong (i.e. empty) scheduler is reconfigured.