Decouple trigger descriptor and scheduler
This uses the Guava event bus to decouple the SQS trigger’s descriptor and the SQS queue monitor scheduler.

Cannot inject the scheduler into the descriptor, since DI is not set up at this time. If Context is turned into an @Extension, DI will be set up, but that results in two instances of the injector which causes other issues.

To resolve this conflict, add a singleton instance of the event bus to communicate configuration changes to the scheduler.